Added migration containing org_type column
Each organisation has a type associated with it. I've added this
as a new type within the database migration when applying it to
the new column org_type - this will stop data that doesn't match
any of the values of the enumerator from being added within this
field. As well as adding this database migration, I've also added
the relevant field to the model to enable this.