Added migration script for project permission columns
This commit adds a migration script to add two new columns
to the projects database. These are permission_type, which
will be used to determine the type of permission selected
by a user that their project requires, and permission_description
which will be used to store details of the permission required
by their project.