Wired up project dashboard functionality (#86)
* Wired up project dashboard functionality

This commit changes the flow of the service journey
to match the current service journey flow diagram.

Upon logging in, a user will be directed to the projects
dashboard page. From here, clicking through to the
Apply for Funding page will initially direct them to create
an organisation if they have not alreaady done so. The end
of the organisation creation journey will direct them
immediately to the create a project journey. For users
that already have an organisation, they will be directed
straight to the create a project journey.

* Removed dashboard test, to be replaced when we pick up unit testing