Added a migration script for capital_work columns to projects table
This commit adds a migrations script containing two new columns
to the projects table. The first, capital_work, is a boolean
column to indicate whether or not capital work will be part
of a project. The second, capital_work_supporting_document
will be used to store an uploaded document, similar to how
the evidence_description column works, but for a single
document.