Implemented fix to nil-value bug on grant request page
This commit fixes a bug which could occur on the 'Your
Grant Request' page if a user had navigated to this page
without setting any project costs or cash contributions.

Previously, the @final_grant_amount instance variable
would have a value of nil, which would cause an error
when the view attempted to use this variable.

Setting .to_i on the values inside the @final_grant_amount
calculation implicitly converts them to 0 if they are nil.

As well, this commit adds an additional message to the
'Your grant request' page in case a @final_grant_amount of 0
is present.