This commit contains refactoring of the organisation legal signatories model, view, and controller. We have refactored the view to use the same nested attributes pattern used elsewhere throughout the service.
Note that a new private method has been included in the controller
which will remove a second legal signatory if it has been added with
empty attributes. We are only required to validate the second legal
signatory if it has been passed through the form with no attributes
populated. This is possible, and has been enabled using a lambda in
the legal_signatory model. However, when ignoring validation, Rails
will attempt to store the record - it's possible to reject a nested
model on the parent model, but it isn't possible to specify the
parameters for this rejection, i.e. only reject the second nested
model passed through. Because of this, we're always choosing not
to validate the second nested model if it hasn't been populated, but
Rails will then save an almost-empty record to the database (auto-
generated attributes are still populated, such as created_at, etc).

The remove_empty_signatory method is therefore necessary to clean
up after ourselves when we have chosen not to validate the second
empty legal signatory.

We have discussed making neither legal signatory mandatory at this
point, depending on how the mandatory version tests. If that is the
case, we can refactor this controller and related models again to
reject all empty objects, something which isn't possible at the
moment, as discussed above.