Added missing fields to organisation summary view
The organisation summary view previously did not contain a row
for the organisation's name, nor did it contain the whole address
for an organisation - only the first line. This commit addresses
both of these issues, as well as tidying up the markup in order
to pass RuboCop line length validation when we bring this in.