Added extra Postgres environment variables to GitHub Action
Our continuous integration workflow GitHub Action started to fail,
reporting an error of 'Failed to initialize, postgres service is unhealthy.'.

Searching for the error found that this seems to be a relatively common
occurence with GitHub Actions, with no reason given as to why. However,
the fix seems to be as simple as adding extra necessary environment
variables to the PostgreSQL service - in this case, we already had
POSTGRES_USER, but were not setting POSTGRES_PASSWORD or POSTGRES_DB.
This commit adds in the missing environment variables, and also sets
the PGPASSWORD environment variable for the 'Run tests' step.

See https://github.community/t5/GitHub-Actions/Mysql-service-never-comes-up-healthy-in-action/td-p/38283.