Refactored routes file and associated references
We had a number of different sections in this file,
some mapped to different models, such as organisation
and project, and others for static pages.

We haven't been consistent with our approach in how
we break out some of these routes into namespaces,
which has left the file quite cluttered and created
some debt.

The debt mostly exists where we haven't been
consistent with our namespacing approach, which has
led to us having to alias routes to then use them
in controllers, tests, etc.

This commit addresses this debt.