Added isAlsoApplicant flag to Salesforce payload
This commit adds the isAlsoApplicant flag to the Salesforce
which is generated from the project model. This flag is used
by Salesforce to determine whether or not to create only
one Contact object if a user and a legal signatory have the
same email address.

As a result, the temporary validation rule we had in place
on the legal signatory model has been removed.