Added validation to alternative finance supporting documents page
This commit adds validation to the onward journey of the alternative
finance supporting documents page. If a user has not uploaded any
supporting documents, then a validation error will be displayed and
the page will be re-rendered. If any supporting documents are
associated with the model object, then choosing to 'Save and
continue' will allow a user to progress to the next page.