Added functionality to strip commas from currency inputs
This commit adds custom setter methods to the GpHefLoan model
which will strip commas from the input before writing the
relevant attribute. This allows us to store the value that a
user entered without having to show an error message based on
commas being entered into the input field on the alternative
finance form.

As well, custom error messages have been added to the locale file
to return a message when a user has entered decimal points into
any of the currency inputs on the alternative finance form.