Refactored file upload handling hooks
This commit refactors the file upload handling hooks, removing
them from their existing partial and moving them into a standalone
JavaScript file. The use of partials for the display of file upload
error messages has been replaced with usage of the existing
JavaScript error-handling functions, which fixes a bug where
duplicate error summaries would appear at the top of the page
when both a file upload and a normal input error occurred on the
same page.

This commit also fixes a bug where errors were not shown if a user
uploaded a file which failed to upload for any reason (such as a
corrupt JSON file).