Added an additional scenario for dashboard_controller spec
This commit adds an additional scenario for the dashboard_controller
spec which ensures that no @funding_applications instance variable
is set if the current user does not have an associated organisation.
As well, the inverse test has an additional line to check that
the @funding_applications instance variable is set appropriately.